Separate mark and sweep!
mark just collects addresses into a set (possibly collecting more
addresses into an existing set).

sweep! walks a sequence (not channel) of addresses, and calls delete-fn
on any address that is both accepted by accept-fn and does not appear
in the marked addresses set.

Removes IGCScratch protocol.

epoch just contains a function-builder-function that returns true if
the timestamp in a key is before an epoch.